Added analytics folder with anaylitics.py and tests
